refactored code broke things up to separate files. cut down bids.js from 300 lines to about 150 lines
