+++ bio-bid/src/utils/enums.js
@@ -1,0 +1,13 @@
+export const Phases = {
+    none: 0,
+    phase1: 1,
+    phase2: 2,
+    phase3: 3
+};
+
+export const Status = {
+  none: 0,
+  open: 1,
+  active: 2,
+  closed: 3
+};
